Restructure the directory to match the library requirement
