--- __init__.py
@@ -1,3 +1,0 @@
-from beans import comment
-from renderer import render_comment_list
-from utils import is_music_available, get_all_music_available
