+++ objection_engine/beans/text.py
